+++ lib/index.ts
@@ -144,41 +144,58 @@ namespace mobservable {
      * }
      */
     export function observable(target:Object, key:string, baseDescriptor?) {
-        // observable annotations are invoked on the prototype, not on actual instances,
+        // - In typescript, observable annotations are invoked on the prototype, not on actual instances,
         // so upon invocation, determine the 'this' instance, and define a property on the
         // instance as well (that hides the propotype property)
-        const isDecoratingProperty = baseDescriptor && !baseDescriptor.hasOwnProperty("value");
-        const descriptor:PropertyDescriptor = baseDescriptor || {};
-        const baseValue = isDecoratingProperty ? descriptor.get : descriptor.value; 
-
-        if (!isDecoratingProperty && typeof baseValue === "function")
-            throw new Error(`@observable functions are deprecated. Use @observable on a getter function if you want to create a view, or wrap the value in 'asReference' if you want to store a value (found on member '${key}').`);
-        if (isDecoratingProperty) {
-            if (typeof baseValue !== "function")
-                throw new Error(`@observable expects a getter function if used on a property (found on member '${key}').`);
-            if (descriptor.set)
-                throw new Error(`@observable properties cannot have a setter (found on member '${key}').`);
-            if (baseValue.length !== 0)
-                throw new Error(`@observable getter functions should not take arguments (found on member '${key}').`);
+        // - In typescript, the baseDescriptor is empty for attributes without initial value
+        // - In babel, the initial value is passed as the closure baseDiscriptor.initializer' 
+        
+        const isDecoratingGetter = baseDescriptor && baseDescriptor.hasOwnProperty("get");
+        const descriptor:PropertyDescriptor = {};
+        let baseValue = undefined;
+        if (baseDescriptor) {
+            if (baseDescriptor.hasOwnProperty('get'))
+                baseValue = baseDescriptor.get;
+            else if (baseDescriptor.hasOwnProperty('value'))
+                baseValue = baseDescriptor.value;
+            else if ((<any>baseDescriptor).initializer) { // For babel
+                baseValue = (<any>baseDescriptor).initializer();
+                if (typeof baseValue === "function")
+                    baseValue = asReference(baseValue);
+            }
         }
-
+    
+        if (!target || typeof target !== "object")
+            throw new Error(`The @observable decorator can only be used on objects`);
+        if (isDecoratingGetter) {
+            if (typeof baseValue !== "function")
+                throw new Error(`@observable expects a getter function if used on a property (in member: '${key}').`);
+            if (descriptor.set)
+                throw new Error(`@observable properties cannot have a setter (in member: '${key}').`);
+            if (baseValue.length !== 0)
+                throw new Error(`@observable getter functions should not take arguments (in member: '${key}').`);
+        }
+    
         descriptor.configurable = true;
         descriptor.enumerable = true;
-        delete descriptor.value;
-        delete descriptor.writable;
         descriptor.get = function() {
-            _.ObservableObject.asReactive(this, null, _.ValueMode.Recursive).set(key, baseValue);
+            // the getter creates a reactive property lazily, so this might even happen during a view.
+            const baseStrict = strict;
+            strict = false;
+            _.ObservableObject.asReactive(this, null,_.ValueMode.Recursive).set(key, baseValue);
+            strict = baseStrict;
             return this[key];
         };
-        descriptor.set = isDecoratingProperty 
+        descriptor.set = isDecoratingGetter 
             ? _.throwingViewSetter
             : function(value) { 
-                _.ObservableObject.asReactive(this, null, _.ValueMode.Recursive).set(key, value); 
+                _.ObservableObject.asReactive(this, null,_.ValueMode.Recursive).set(key, typeof value === "function" ? asReference(value) : value); 
             }
         ;
-
-        if (!isDecoratingProperty) {
-            Object.defineProperty(target, key, descriptor);
+        if (!baseDescriptor) {
+            Object.defineProperty(target, key, descriptor); // For typescript
+        } else { 
+            return descriptor;
         }
     }